separated user role and coordinate converters
